# src/main/resources/api.yml
openapi: 3.0.0
info:
  title: Campus Wall API
  version: "1.0.0"
  description: API for campus social features
servers:
  - url: /api/v1
...
paths:
...
  /users/me/comments:
    get:
      tags: [users]
      summary: Get current user's own comments
      description: Retrieve all comments made by the authenticated user with pagination and sorting
      operationId: getUserComments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [NEWEST, OLDEST]
            default: NEWEST
          description: Sort order for comments
      responses:
        "200":
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CommentDTO"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        "401":
          description: Unauthorized - authentication required
  /users/me/posts:
    get:
      tags: [users]
      summary: Get current user's own posts
      description: Retrieve all posts created by the authenticated user with pagination and sorting
      operationId: getUserPosts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [NEWEST, OLDEST, MOST_LIKED, LEAST_LIKED]
            default: NEWEST
          description: Sort order for posts
      responses:
        "200":
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/PostDTO"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        "401":
          description: Unauthorized - authentication required
  /users/me/internships:
    get:
      tags: [users]
      summary: Get current user's own internships
      description: Retrieve all internship postings created by the authenticated user with pagination and sorting
      operationId: getUserInternships
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [NEWEST, OLDEST]
            default: NEWEST
          description: Sort order for internships
      responses:
        "200":
          description: Internships retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/InternshipDTO"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        "401":
          description: Unauthorized - authentication required
  /users/me/marketplaces:
    get:
      tags: [users]
      summary: Get current user's own marketplace items
      description: Retrieve all marketplace items listed by the authenticated user with pagination and sorting
      operationId: getUserMarketplaces
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: sort
          required: false
          schema:
            type: string
            enum: [NEWEST, OLDEST]
            default: NEWEST
          description: Sort order for marketplace items
      responses:
        "200":
          description: Marketplace items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ItemDTO"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
        "401":
          description: Unauthorized - authentication required
...
  /admin/users:
    get:
      tags: [admin]
      summary: Get all users
      description: Admin endpoint to retrieve all users with pagination, filtering, and sorting
      operationId: adminGetUsers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: blocked
          required: false
          schema:
            type: boolean
          description: Filter by blocked status (true/false)
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
            enum: [createdAt, schoolDomain, reportCount]
          description: Field to sort by
        - in: query
          name: sortOrder
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (ascending or descending)
      responses:
        "200":
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUserDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
  /admin/users/{id}:
    get:
      tags: [admin]
      summary: Get user by ID
      description: Admin endpoint to retrieve specific user details
      operationId: adminGetUserById
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
      responses:
        "200":
          description: User retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDTO"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: User not found
  /admin/users/{id}/block:
    put:
      tags: [admin]
      summary: Block a user
      description: Admin endpoint to block a user from logging in and posting
      operationId: adminBlockUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID to block
      responses:
        "200":
          description: User blocked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User blocked successfully"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: User not found
  /admin/users/{id}/unblock:
    put:
      tags: [admin]
      summary: Unblock a user
      description: Admin endpoint to unblock a previously blocked user
      operationId: adminUnblockUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID to unblock
      responses:
        "200":
          description: User unblocked successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User unblocked successfully"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: User not found
  /admin/users/{id}/posts:
    get:
      tags: [admin]
      summary: Get all posts by a user
      description: Admin endpoint to retrieve all posts created by a specific user with pagination and sorting
      operationId: adminGetUserPosts
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
            enum: [createdAt, likeCount, commentCount, userId]
          description: Field to sort by
        - in: query
          name: sortOrder
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (ascending or descending)
      responses:
        "200":
          description: User posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminPostDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: User not found
  /admin/users/{id}/comments:
    get:
      tags: [admin]
      summary: Get all comments by a user
      description: Admin endpoint to retrieve all comments created by a specific user with pagination and sorting
      operationId: adminGetUserComments
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: User ID
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number for pagination
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of results per page
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
            enum: [createdAt, userId]
          description: Field to sort by
        - in: query
          name: sortOrder
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order (ascending or descending)
      responses:
        "200":
          description: User comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminCommentDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized - authentication required
        "403":
          description: Forbidden - admin access required
        "404":
          description: User not found
  /admin/users/{id}/internships:
    get:
      tags: [admin]
      summary: Get all internships by a user
      operationId: adminGetUserInternships
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: User internships retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminInternshipDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users/{id}/marketplaces:
    get:
      tags: [admin]
      summary: Get all marketplace items by a user
      operationId: adminGetUserMarketplaces
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: User marketplace items retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminMarketplaceDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users/{id}/conversations:
    get:
      tags: [admin]
      summary: Get all conversations for a user
      operationId: adminGetUserConversations
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: User conversations retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminConversationDTO"
                  pagination:
                    $ref: "#/components/schemas/PaginationInfo"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
...
components:
  schemas:
...
    PostDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          description: Post title
        content:
          type: string
        wall:
          type: string
          enum: [campus, national]
          description: Which wall the post is on
        author:
          type: object
          properties:
            id:
              type: string
            profileName:
              type: string
              description: Profile name of the post author
            isAnonymous:
              type: boolean
              description: Whether author is anonymous
        likes:
          type: integer
          description: Number of likes
        comments:
          type: integer
          description: Number of comments
        liked:
          type: boolean
          description: Whether current user has liked this post
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CommentDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
          description: ID of the parent entity (post, internship, or marketplace item)
        parentType:
          type: string
          enum: [POST, INTERNSHIP, MARKETPLACE]
          description: Type of the parent entity
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        author:
          type: object
          properties:
            id:
              type: string
            profileName:
              type: string
              description: Profile name of the comment author
            isAnonymous:
              type: boolean

    # ----- Internships -----
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT